+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-## [UNRELEASED] neptune-prophet 0.0.1
+## [UNRELEASED] neptune-prophet 0.2.0
+
+### Fixes
+- Fixed typo in length error ([#8](https://github.com/neptune-ai/neptune-prophet/pull/8))
+- Added missing file extension for serialized model ([#8](https://github.com/neptune-ai/neptune-prophet/pull/8))
+
+## neptune-prophet 0.1.0
 
 ### Features
 - Logging model metadata, serialized models, Prophet's plots, and additional diagnostic plots ([#3](https://github.com/neptune-ai/neptune-prophet/pull/3))